pyproject.toml
[project]
name = "enana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "sphinx>=8.2.3",
    "sphinx-markdown-builder>=0.6.8",
    "sphinx-rtd-theme>=3.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "black>=25.11.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["enana*"]

[tool.black]
line-length = 79
target-version = ['py313']
